+++ src/RequestOpts.php
@@ -1,0 +1,59 @@
+<?php
+/**
+ * Created by PhpStorm.
+ * User: lkaemmerling
+ * Date: 05.09.18
+ * Time: 10:55
+ */
+
+namespace LKDev\HetznerCloud;
+
+
+/**
+ * Class RequestOpts
+ * @package LKDev\HetznerCloud
+ */
+class RequestOpts
+{
+
+    /**
+     * @var int
+     */
+    public $perPage;
+
+
+    /**
+     * @var int
+     */
+    public $page;
+
+
+    /**
+     * @var string
+     */
+    public $labelSelector;
+
+    /**
+     * RequestOpts constructor.
+     * @param $perPage
+     * @param $page
+     * @param $labelSelector
+     */
+    public function __construct(int $perPage = null, int $page = null, string $labelSelector = null)
+    {
+        $this->perPage = $perPage;
+        $this->page = $page;
+        $this->labelSelector = $labelSelector;
+    }
+
+    /**
+     * @return string
+     */
+    public function buildQuery()
+    {
+        $values = collect(get_object_vars($this))->reject(function ($var) {
+            return $var == null;
+        })->toArray();
+        return count($values) > 0 ? '': ('?'.http_build_query($values));
+    }
+}